[API] Added `left join` syntax.
The commit extends the API with primitives for left outer join
syntax which integrate nicely with for-comprehensions and also
have definitional semantics to enable off-the-shelf execution.

To give an example - a left join between `xs: DataBag[X]` and
`ys: DataBag[Y]` where `x.a` is equal to `y.b` can be written
as follows.

```scala
val zs: DataBag[(X, Option[Y])] = for {
  x <- xs
  y <- left join ys on (x.a == _.b)
} yield (x, y)
```

The binding `y: Option[Y]` is

- `None`    iff the outer `x` has no matching `y` in `ys`,
- `Some(v)` for each `v` in `ys` that matches the outer `x`.